Fixed `max_age` in settings to use int value.

+++ web/sites/default/includes/modules/settings.system.php
@@ -8,7 +8,7 @@
 declare(strict_types=1);
 
 // Expiration of cached pages.
-$config['system.performance']['cache']['page']['max_age'] = getenv('DRUPAL_CACHE_PAGE_MAX_AGE') ?: 900;
+$config['system.performance']['cache']['page']['max_age'] = (int) (getenv('DRUPAL_CACHE_PAGE_MAX_AGE') ?: 900);
 
 if ($settings['environment'] === ENVIRONMENT_PROD) {
   // Always aggregate CSS and JS files in production.
